+++ flink/flink-1.14/src/test/java/io/tidb/bigdata/flink/tidb/pushdown/FilterPushDownNotEqualTest.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package io.tidb.bigdata.flink.tidb;
+package io.tidb.bigdata.flink.tidb.pushdown;
 
-import static io.tidb.bigdata.flink.tidb.FilterPushDownValidator.doTestFilter;
-import static io.tidb.bigdata.flink.tidb.FilterPushDownValidator.getColumnType;
-import static io.tidb.bigdata.flink.tidb.FilterPushDownValidator.rows;
+import static io.tidb.bigdata.flink.tidb.pushdown.FilterPushDownValidator.doTestFilter;
+import static io.tidb.bigdata.flink.tidb.pushdown.FilterPushDownValidator.getColumnType;
+import static io.tidb.bigdata.flink.tidb.pushdown.FilterPushDownValidator.rows;
 
 import com.google.common.collect.ImmutableList;
 import io.tidb.bigdata.test.IntegrationTest;
@@ -36,7 +36,7 @@ import org.tikv.common.types.DataType;
 import org.tikv.common.types.StringType;
 
 @Category(IntegrationTest.class)
-public class FilterPushDownLessThanTest {
+public class FilterPushDownNotEqualTest {
 
   /**
    * Filters shot will return correct rows, and filters missed will return empty row list.
@@ -48,208 +48,209 @@ public class FilterPushDownLessThanTest {
     String column = "c1";
     Object value = 2;
     DataType type = getColumnType(column);
-    Expression expression = Expressions.lessThan(Expressions.column(column, type),
+    Expression expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < %s", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> %s", column, value));
     value = 1;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < %s", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> %s", column, value));
     // Swap column and value, we only test for tinyint, other types are same as tinyint
-    value = 0;
-    expression = Expressions.lessThan(Expressions.constant(value, type),
+    value = 2;
+    expression = Expressions.notEqual(Expressions.constant(value, type),
         Expressions.column(column, type));
-    doTestFilter(rows, expression, String.format("%s < `%s`", value, column));
+    doTestFilter(rows, expression, String.format("%s <> `%s`", value, column));
     value = 1;
-    expression = Expressions.lessThan(Expressions.constant(value, type),
+    expression = Expressions.notEqual(Expressions.constant(value, type),
         Expressions.column(column, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("%s < `%s`", value, column));
+    doTestFilter(ImmutableList.of(), expression, String.format("%s <> `%s`", value, column));
 
     // smallint
     column = "c2";
     value = 2;
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < %s", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> %s", column, value));
     value = 1;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < %s", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> %s", column, value));
 
     // mediumint
     column = "c3";
     value = 2;
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < %s", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> %s", column, value));
     value = 1;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < %s", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> %s", column, value));
 
     // int
     column = "c4";
     value = 2;
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < %s", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> %s", column, value));
     value = 1;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < %s", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> %s", column, value));
 
     // bigint
     column = "c5";
     value = 2;
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < %s", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> %s", column, value));
     value = 1;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < %s", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> %s", column, value));
 
     // char
     column = "c6";
     value = "chartype1";
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> '%s'", column, value));
     value = "chartype";
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> '%s'", column, value));
 
     // varchar
     column = "c7";
     value = "varchartype1";
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> '%s'", column, value));
     value = "varchartype";
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> '%s'", column, value));
 
     // tinytext
     column = "c8";
     value = "tinytexttype1";
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> '%s'", column, value));
     value = "tinytexttype";
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> '%s'", column, value));
 
     // mediumtext
     column = "c9";
     value = "mediumtexttype1";
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> '%s'", column, value));
     value = "mediumtexttype";
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> '%s'", column, value));
 
     // text
     column = "c10";
     value = "texttype1";
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> '%s'", column, value));
     value = "texttype";
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> '%s'", column, value));
 
     // longtext
     column = "c11";
     value = "longtexttype1";
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> '%s'", column, value));
     value = "longtexttype";
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(ImmutableList.of(), expression, String.format("`%s` < '%s'", column, value));
+    doTestFilter(ImmutableList.of(), expression, String.format("`%s` <> '%s'", column, value));
 
     // float
     column = "c18";
     value = 1.235;
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < CAST(%s AS FLOAT)", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> CAST(%s AS FLOAT)", column, value));
     value = 1.234;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(ImmutableList.of(), expression,
-        String.format("`%s` < CAST(%s AS FLOAT)", column, value));
+        String.format("`%s` <> CAST(%s AS FLOAT)", column, value));
 
     // double
     column = "c19";
-    value = 2.4567891;
+    value = 2.456789123;
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < CAST(%s AS DOUBLE)", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> CAST(%s AS DOUBLE)", column, value));
     value = 2.456789;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(ImmutableList.of(), expression,
-        String.format("`%s` < CAST(%s AS DOUBLE)", column, value));
+        String.format("`%s` <> CAST(%s AS DOUBLE)", column, value));
 
     // decimal
     column = "c20";
-    value = 123.457;
+    value = 123.4567;
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < CAST(%s AS DECIMAL(6,3))", column, value));
+    doTestFilter(rows, expression,
+        String.format("`%s` <> CAST(%s AS DECIMAL(6,3))", column, value));
     value = 123.456;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(ImmutableList.of(), expression,
-        String.format("`%s` < CAST(%s AS DECIMAL(6,3))", column, value));
+        String.format("`%s` <> CAST(%s AS DECIMAL(6,3))", column, value));
 
     // date
     column = "c21";
     value = Date.valueOf("2020-08-11");
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
-    doTestFilter(rows, expression, String.format("`%s` < CAST('%s' AS DATE)", column, value));
+    doTestFilter(rows, expression, String.format("`%s` <> CAST('%s' AS DATE)", column, value));
     value = Date.valueOf("2020-08-10");
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(ImmutableList.of(), expression,
-        String.format("`%s` < CAST('%s' AS DATE)", column, value));
+        String.format("`%s` <> CAST('%s' AS DATE)", column, value));
 
     // datetime
     column = "c23";
     value = Timestamp.valueOf("2020-08-10 15:30:30");
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(rows, expression,
-        String.format("`%s` < CAST('%s' AS TIMESTAMP(6))", column, value));
+        String.format("`%s` <> CAST('%s' AS TIMESTAMP(6))", column, value));
     value = Timestamp.valueOf("2020-08-10 15:30:29");
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(ImmutableList.of(), expression,
-        String.format("`%s` < CAST('%s' AS TIMESTAMP(6))", column, value));
+        String.format("`%s` <> CAST('%s' AS TIMESTAMP(6))", column, value));
 
     // timestamp
     column = "c24";
@@ -259,48 +260,48 @@ public class FilterPushDownLessThanTest {
     ZonedDateTime utc = zonedDateTime.withZoneSameInstant(ZoneId.of("UTC"));
     Timestamp utcTimestamp = Timestamp.valueOf(utc.toLocalDateTime());
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(utcTimestamp, type));
     doTestFilter(rows, expression,
-        String.format("`%s` < CAST('%s' AS TIMESTAMP(6))", column, timestamp));
+        String.format("`%s` <> CAST('%s' AS TIMESTAMP(6))", column, timestamp));
 
     timestamp = Timestamp.valueOf("2020-08-10 16:30:29");
     zonedDateTime = ZonedDateTime.of(timestamp.toLocalDateTime(),
         ZoneId.systemDefault());
     utc = zonedDateTime.withZoneSameInstant(ZoneId.of("UTC"));
     utcTimestamp = Timestamp.valueOf(utc.toLocalDateTime());
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(utcTimestamp, type));
     doTestFilter(ImmutableList.of(), expression,
-        String.format("`%s` < CAST('%s' AS TIMESTAMP(6))", column, timestamp));
+        String.format("`%s` <> CAST('%s' AS TIMESTAMP(6))", column, timestamp));
 
     // year
     column = "c25";
     value = 2021;
     type = getColumnType(column);
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(rows, expression,
-        String.format("`%s` < %s", column, value));
+        String.format("`%s` <> %s", column, value));
     value = 2020;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(ImmutableList.of(), expression,
-        String.format("`%s` < %s", column, value));
+        String.format("`%s` <> %s", column, value));
 
     // enum
     column = "c28";
     value = "2";
     type = StringType.VARCHAR;
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(rows, expression,
-        String.format("`%s` < '%s'", column, value));
+        String.format("`%s` <> '%s'", column, value));
     value = "1";
-    expression = Expressions.lessThan(Expressions.column(column, type),
+    expression = Expressions.notEqual(Expressions.column(column, type),
         Expressions.constant(value, type));
     doTestFilter(ImmutableList.of(), expression,
-        String.format("`%s` < '%s'", column, value));
+        String.format("`%s` <> '%s'", column, value));
   }
 
   /**
@@ -313,15 +314,15 @@ public class FilterPushDownLessThanTest {
     // json
     String column = "c27";
     Object value = "{\"a\": 1, \"b\": 2}";
-    doTestFilter(rows, null, String.format("`%s` < '%s'", column, value));
+    doTestFilter(rows, null, String.format("`%s` <> '%s'", column, value));
 
     // set
     column = "c29";
     value = "a";
-    doTestFilter(rows, null, String.format("`%s` < '%s'", column, value));
+    doTestFilter(rows, null, String.format("`%s` <> '%s'", column, value));
 
     // columns to columns
-    doTestFilter(rows, null, "`c1` < `c2`");
+    doTestFilter(rows, null, "`c1` <> `c2`");
   }
 
 }
